+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - cli: Added the `treasury batch-withdraw` subcommand.
+- cli: Added `--authority` option for the `inspect price-feed` subcommand.
+- cli: Introduced a new ALT type `PriceFeed` for the `alt extend` subcommand.
 
 ### Changed